updates model factory
